# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Scientific Publication Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface RegisterInput {
  email: string;
  /** @minLength 6 */
  password: string;
  name: string;
  /** @nullable */
  affiliation?: string | null;
  /** @nullable */
  orcidId?: string | null;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  USER: "USER",
  VERIFIER: "VERIFIER",
  ADMIN: "ADMIN",
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  /** @nullable */
  verifierLayer?: number | null;
  /** @nullable */
  affiliation?: string | null;
  /** @nullable */
  orcidId?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type RoleUpdateRole =
  (typeof RoleUpdateRole)[keyof typeof RoleUpdateRole];

export const RoleUpdateRole = {
  USER: "USER",
  VERIFIER: "VERIFIER",
  ADMIN: "ADMIN",
} as const;

export interface RoleUpdate {
  role: RoleUpdateRole;
  /** @nullable */
  verifierLayer?: number | null;
}

export interface PaperInput {
  title: string;
  abstract: string;
  keywords: string[];
  category: string;
  pdfUrl: string;
  content: string;
  coAuthors?: string[];
}

export interface PaperUpdate {
  title?: string;
  abstract?: string;
  keywords?: string[];
  category?: string;
  pdfUrl?: string;
  content?: string;
  coAuthors?: string[];
}

export type PaperSummaryStatus =
  (typeof PaperSummaryStatus)[keyof typeof PaperSummaryStatus];

export const PaperSummaryStatus = {
  DRAFT: "DRAFT",
  SUBMITTED: "SUBMITTED",
  AI_REVIEW: "AI_REVIEW",
  AI_PASSED: "AI_PASSED",
  AI_FAILED: "AI_FAILED",
  LAYER_2_REVIEW: "LAYER_2_REVIEW",
  LAYER_2_APPROVED: "LAYER_2_APPROVED",
  LAYER_3_REVIEW: "LAYER_3_REVIEW",
  LAYER_3_APPROVED: "LAYER_3_APPROVED",
  PUBLISHED: "PUBLISHED",
  REVISION_REQUESTED: "REVISION_REQUESTED",
  REJECTED: "REJECTED",
} as const;

export interface PaperSummary {
  id: number;
  title: string;
  abstract?: string;
  keywords?: string[];
  category: string;
  status: PaperSummaryStatus;
  currentLayer?: number;
  /** @nullable */
  doi?: string | null;
  authorId: number;
  authorName: string;
  pdfUrl?: string;
  /** @nullable */
  publishedAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type AiIssueSeverity =
  (typeof AiIssueSeverity)[keyof typeof AiIssueSeverity];

export const AiIssueSeverity = {
  critical: "critical",
  warning: "warning",
  info: "info",
} as const;

export interface AiIssue {
  type: string;
  severity: AiIssueSeverity;
  message: string;
  /** @nullable */
  location?: string | null;
}

export interface AiReport {
  score: number;
  passed: boolean;
  wordCount?: number;
  aiGeneratedLikelihood?: number;
  plagiarismScore?: number;
  structureComplete?: boolean;
  citationValid?: boolean;
  grammarScore?: number;
  issues?: AiIssue[];
}

export type ReviewDecision =
  (typeof ReviewDecision)[keyof typeof ReviewDecision];

export const ReviewDecision = {
  APPROVED: "APPROVED",
  REJECTED: "REJECTED",
  REVISION: "REVISION",
} as const;

export interface InlineNote {
  page?: number;
  position?: string;
  note: string;
}

export interface Review {
  id: number;
  paperId: number;
  reviewerId: number;
  reviewerName: string;
  layer: number;
  decision: ReviewDecision;
  comments: string;
  inlineNotes?: InlineNote[];
  createdAt: string;
}

export interface StatusHistoryEntry {
  id: number;
  /** @nullable */
  fromStatus: string | null;
  toStatus: string;
  changedBy: string;
  /** @nullable */
  note?: string | null;
  createdAt: string;
}

export type PaperDetail = PaperSummary & {
  aiReport?: AiReport;
  reviews?: Review[];
  statusHistory?: StatusHistoryEntry[];
  /** @nullable */
  assignedVerifierId?: number | null;
  /** @nullable */
  assignedVerifierName?: string | null;
  coAuthors?: string[];
};

export interface PaperListResponse {
  papers: PaperSummary[];
  total: number;
  page: number;
  limit: number;
}

export type ReviewInputDecision =
  (typeof ReviewInputDecision)[keyof typeof ReviewInputDecision];

export const ReviewInputDecision = {
  APPROVED: "APPROVED",
  REJECTED: "REJECTED",
  REVISION: "REVISION",
} as const;

export interface ReviewInput {
  decision: ReviewInputDecision;
  comments: string;
  inlineNotes?: InlineNote[];
}

export interface AssignInput {
  verifierId: number;
}

export type StatusOverrideStatus =
  (typeof StatusOverrideStatus)[keyof typeof StatusOverrideStatus];

export const StatusOverrideStatus = {
  DRAFT: "DRAFT",
  SUBMITTED: "SUBMITTED",
  AI_REVIEW: "AI_REVIEW",
  AI_PASSED: "AI_PASSED",
  AI_FAILED: "AI_FAILED",
  LAYER_2_REVIEW: "LAYER_2_REVIEW",
  LAYER_2_APPROVED: "LAYER_2_APPROVED",
  LAYER_3_REVIEW: "LAYER_3_REVIEW",
  LAYER_3_APPROVED: "LAYER_3_APPROVED",
  PUBLISHED: "PUBLISHED",
  REVISION_REQUESTED: "REVISION_REQUESTED",
  REJECTED: "REJECTED",
} as const;

export interface StatusOverride {
  status: StatusOverrideStatus;
  note?: string;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface PlatformStats {
  totalPublished: number;
  totalSubmitted: number;
  totalUsers: number;
  categories: CategoryCount[];
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface TimeSeriesPoint {
  date: string;
  count: number;
}

export interface AnalyticsData {
  statusBreakdown: StatusCount[];
  submissionsOverTime: TimeSeriesPoint[];
  avgReviewTime: number;
  acceptanceRate: number;
  totalPapers: number;
  totalUsers: number;
}

export type ListPublishedPapersParams = {
  search?: string;
  category?: string;
  page?: number;
  limit?: number;
};

export type AdminListPapersParams = {
  status?: string;
  page?: number;
  limit?: number;
};