# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Scientific Publication Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Papers (public) ─────────────────────────────────────────────────────────
...
  /admin/papers:
    get:
      operationId: adminListPapers
      tags: [admin]
      summary: List all papers with full status info
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: All papers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperListResponse"
  /admin/papers/{id}/assign:
    post:
      operationId: assignVerifier
      tags: [admin]
      summary: Assign a verifier to a paper
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignInput"
      responses:
        "200":
          description: Verifier assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperDetail"
  /admin/papers/{id}/status:
    patch:
      operationId: adminUpdatePaperStatus
      tags: [admin]
      summary: Override paper status (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusOverride"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaperDetail"
...
  /admin/users/{id}/role:
    patch:
      operationId: adminUpdateUserRole
      tags: [admin]
      summary: Update a user's role
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpdate"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
        affiliation:
          type: ["string", "null"]
        orcidId:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string

    # ─── User ──────────────────────────────────────────────────────────────────
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [USER, VERIFIER, ADMIN]
        verifierLayer:
          type: ["integer", "null"]
        affiliation:
          type: ["string", "null"]
        orcidId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    RoleUpdate:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [USER, VERIFIER, ADMIN]
        verifierLayer:
          type: ["integer", "null"]

    # ─── Paper ─────────────────────────────────────────────────────────────────
...
    PaperSummary:
      type: object
      required: [id, title, status, category, authorId, authorName, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        abstract:
          type: string
        keywords:
          type: array
          items:
            type: string
        category:
          type: string
        status:
          type: string
          enum: [DRAFT, SUBMITTED, AI_REVIEW, AI_PASSED, AI_FAILED, LAYER_2_REVIEW, LAYER_2_APPROVED, LAYER_3_REVIEW, LAYER_3_APPROVED, PUBLISHED, REVISION_REQUESTED, REJECTED]
        currentLayer:
          type: integer
        doi:
          type: ["string", "null"]
        authorId:
          type: integer
        authorName:
          type: string
        pdfUrl:
          type: string
        publishedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PaperDetail:
      allOf:
        - $ref: "#/components/schemas/PaperSummary"
        - type: object
          properties:
            aiReport:
              $ref: "#/components/schemas/AiReport"
            reviews:
              type: array
              items:
                $ref: "#/components/schemas/Review"
            statusHistory:
              type: array
              items:
                $ref: "#/components/schemas/StatusHistoryEntry"
            assignedVerifierId:
              type: ["integer", "null"]
            assignedVerifierName:
              type: ["string", "null"]
            coAuthors:
              type: array
              items:
                type: string
    PaperListResponse:
      type: object
      required: [papers, total, page, limit]
      properties:
        papers:
          type: array
          items:
            $ref: "#/components/schemas/PaperSummary"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ─── AI Report ─────────────────────────────────────────────────────────────
    AiReport:
      type: object
      required: [score, passed]
      properties:
        score:
          type: number
        passed:
          type: boolean
        wordCount:
          type: integer
        aiGeneratedLikelihood:
          type: number
        plagiarismScore:
          type: number
        structureComplete:
          type: boolean
        citationValid:
          type: boolean
        grammarScore:
          type: number
        issues:
          type: array
          items:
            $ref: "#/components/schemas/AiIssue"
    AiIssue:
      type: object
      required: [type, severity, message]
      properties:
        type:
          type: string
        severity:
          type: string
          enum: [critical, warning, info]
        message:
          type: string
        location:
          type: ["string", "null"]

    # ─── Review ────────────────────────────────────────────────────────────────
...
    Review:
      type: object
      required: [id, paperId, reviewerId, reviewerName, layer, decision, comments, createdAt]
      properties:
        id:
          type: integer
        paperId:
          type: integer
        reviewerId:
          type: integer
        reviewerName:
          type: string
        layer:
          type: integer
        decision:
          type: string
          enum: [APPROVED, REJECTED, REVISION]
        comments:
          type: string
        inlineNotes:
          type: array
          items:
            $ref: "#/components/schemas/InlineNote"
        createdAt:
          type: string
          format: date-time
    InlineNote:
      type: object
      required: [note]
      properties:
        page:
          type: integer
        position:
          type: string
        note:
          type: string

    # ─── StatusHistory ─────────────────────────────────────────────────────────
    StatusHistoryEntry:
      type: object
      required: [id, fromStatus, toStatus, changedBy, createdAt]
      properties:
        id:
          type: integer
        fromStatus:
          type: ["string", "null"]
        toStatus:
          type: string
        changedBy:
          type: string
        note:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    # ─── Admin ops ─────────────────────────────────────────────────────────────
    AssignInput:
      type: object
      required: [verifierId]
      properties:
        verifierId:
          type: integer
    StatusOverride:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [DRAFT, SUBMITTED, AI_REVIEW, AI_PASSED, AI_FAILED, LAYER_2_REVIEW, LAYER_2_APPROVED, LAYER_3_REVIEW, LAYER_3_APPROVED, PUBLISHED, REVISION_REQUESTED, REJECTED]
        note:
          type: string

    # ─── Analytics ─────────────────────────────────────────────────────────────
...