# lib/api-zod/src/generated/types/aiReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Scientific Publication Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AiIssue } from "./aiIssue";

export interface AiReport {
  score: number;
  passed: boolean;
  wordCount?: number;
  aiGeneratedLikelihood?: number;
  plagiarismScore?: number;
  structureComplete?: boolean;
  citationValid?: boolean;
  grammarScore?: number;
  issues?: AiIssue[];
}