# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Scientific Publication Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  /** @nullable */
  verifierLayer?: number | null;
  /** @nullable */
  affiliation?: string | null;
  /** @nullable */
  orcidId?: string | null;
  createdAt: Date;
}